requirements.txt
these requirement show install to the run the server

to insall all of the following packages simply run the command on your terminal
pip install -r requirements.txt(means fille)

if some one not give such a requirement.txt file no worry about that simply run the command
pip insall fastapi[all]


alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
env-file==2020.12.3
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
httptools==0.2.0
idna==3.3
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mysql==0.0.3
mysqlclient==2.1.0
oauth2==1.9.0.post1
orjson==3.6.7
passlib==1.7.4
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
toml==0.10.2
typing-extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
values==2020.12.3
watchgod==0.7
websockets==10.2
zipp==3.7.0
